twitter_producer/requirements.txt
#
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
kafka-python
tweepy